Refactoring code
Created new "utils" folder and moved the following items to it:
- src/common/hooks/useFetch.ts
- src/common/generateRandomId.ts
- src/common/Settings.ts
